\docType{data}
\name{bowhead.hmm.pars.bs}
\alias{bowhead.hmm.pars.bs}
\title{Bootstrapped Greenland bowhead whale availability HMM parameters.}
\format{A vector of parameters created using \code{\link{vectorize.hmmpars}}.}
\description{
  Hidden Markov model (HMM) parameter estimates for bowhead
  whale availability obtained from fitting HMMs (using
  library \code{\link{HiddenMarkov}}) to parametric
  bootstrap samples of availability data
  \code{\link{bowhead.hmm.pars}}. *** Need to complete the
  documentatoin ***.
}
\keyword{datasets}

