\name{calc.derived}
\alias{calc.derived}
\title{Calculates a bunch of derived statistics from model.}
\usage{
  calc.derived(stat, hmmlt, obs = 1:dim(hmmlt$xy)[1])
}
\arguments{
  \item{stat.}{}

  \item{hmmlt.}{}

  \item{obs.}{}
}
\description{
  Calculates a bunch of derived statistics from model.
}

