% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cv.avail}
\alias{cv.avail}
\title{Calculate CV of mean time available.}
\usage{
cv.avail(hhm.pars, B = 1000)
}
\arguments{
\item{B}{number of bootstrap replicates to do.}

\item{hmm.pars}{object of class \code{hmm.pars}.}
}
\description{
Calculates CV of mean time available and unavailable from an hmm.pars object.
}

