% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/derived_stats.R
\name{fitted_invesw}
\alias{fitted_invesw}
\title{Calculates 1/esw from model.}
\usage{
fitted_invesw(hmmlt, obs = 1:dim(hmmlt$xy)[1], nx = 100)
}
\arguments{
\item{hmmlt}{output from \code{\link{fit.hmltm}}}

\item{obs}{observations (row numbers of \code{hmmlt$xy}) for which to calculate esw}

\item{nx}{number of x-values (perpendicular distance values) to use in calculation.}
}
\description{
Calculates inverse of effective strip half-width, 1/(W*E[p(x)]) (where W is actual half-width) for
given observations, from model.
}

