% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hsltm.R
\name{fitted_p}
\alias{fitted_p}
\title{Calculates E[p(x)] from model.}
\usage{
fitted_p(hmmlt, obs = 1:dim(hmmlt$xy)[1], nx = 100)
}
\arguments{
\item{hmmlt}{output from \code{\link{fit.hmltm}}}

\item{obs}{observations (row numbers of \code{hmmlt$xy}) for which to calculate esw}

\item{nx}{number of x-values (perpendicular distance values) to use in calculation.}
}
\description{
Calculates mean over perpendicular distance (x) of fitted values, p(x) for given observations,
from model.
}

