% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hsltm.R
\name{gety.obs}
\alias{gety.obs}
\title{Return all (discrete) forward distances in veiw.}
\usage{
gety.obs(Ix, Iyobs, Iymax, Idy)
}
\arguments{
\item{Ix}{Perpendicular distance.}

\item{Iyobs}{Forward distance.}

\item{Iymax}{Maximum forward distance in view.}

\item{Idy}{Forward distance step increment.}
}
\description{
C++ function to return all (discrete) forward distances in veiw.
}

