% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gety_obs_rcpp}
\alias{gety_obs_rcpp}
\title{Forward distances in view}
\usage{
gety_obs_rcpp(x, null_yobs, yobs, ymax, dy)
}
\arguments{
\item{x}{Perpendicular distance}

\item{null_yobs}{???}

\item{yobs}{Forward distance}

\item{ymax}{Maximum forward distance in view}

\item{dy}{Forward distance step increment}
}
\value{
A vector of the forward distances in view.
}
\description{
Computes all discrete forward distances in view.
Written in C++.
}

