% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{h_rcpp}
\alias{h_rcpp}
\title{Hazard detection function}
\usage{
h_rcpp(x, y, b, hfun)
}
\arguments{
\item{x}{Perpendicular distance of the detection}

\item{y}{Forward distance of the detection}

\item{b}{Parameters of the hazard function}

\item{hfun}{Name of hazard detection function}
}
\value{
The value of the hazard function at (x,y)
}
\description{
Used within the C++ code, and in the simulation function simData.
}

