\name{hmltm.invp}
\alias{hmltm.invp}
\title{Calculates 1/esw from model.}
\usage{
  hmltm.invp(pars, hfun, models = list(y = NULL, x = NULL),
    cov = NULL, survey.pars, hmm.pars, nx = 100,
    type = "response")
}
\arguments{
  \item{pars.}{}

  \item{hfun.}{}

  \item{models.}{}

  \item{cov.}{}

  \item{survey.pars.}{}

  \item{hmm.pars.}{}

  \item{nx.}{}

  \item{type.}{}
}
\description{
  Calculates 1/esw from model.
}

