\name{hmltm.p}
\alias{hmltm.p}
\title{Calculates E[p] from model.}
\usage{
  hmltm.p(pars, hfun, models = list(y = NULL, x = NULL),
    cov = NULL, survey.pars, hmm.pars, nx = 100,
    type = "response")
}
\arguments{
  \item{pars.}{}

  \item{hfun.}{}

  \item{models.}{}

  \item{cov.}{}

  \item{survey.pars.}{}

  \item{hmm.pars.}{}

  \item{nx.}{}

  \item{type.}{}
}
\description{
  Calculates E[p] from model.
}

