\name{hmltm.px}
\alias{hmltm.px}
\title{Calculates perp dist detection probability p(x).}
\usage{
  hmltm.px(x, pars, hfun,
    models = list(y = NULL, x = NULL), cov = NULL,
    survey.pars, hmm.pars, ny = 100, type = "response",
    ally = TRUE)
}
\arguments{
  \item{x.}{}

  \item{pars.}{}

  \item{hfun.}{}

  \item{models.}{}

  \item{cov.}{}

  \item{survey.pars.}{}

  \item{hmm.pars.}{}

  \item{ny.}{}
}
\description{
  Calculates perp dist detection probability p(x).
}

