\name{hmltm.stat}
\alias{hmltm.stat}
\title{Calculates various statistics from model.}
\usage{
  hmltm.stat(stat, b, hfun,
    models = list(y = NULL, x = NULL), cov = NULL,
    survey.pars, hmm.pars, nx = 100, type = "link")
}
\arguments{
  \item{stat.}{}

  \item{b.}{}

  \item{hfun.}{}

  \item{models.}{}

  \item{cov.}{}

  \item{survey.pars.}{}

  \item{hmm.pars.}{}

  \item{nx.}{}

  \item{type.}{}
}
\description{
  Calculates various statistics from model.
}

