% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{n2w_rcpp}
\alias{n2w_rcpp}
\title{Parameters transformation}
\usage{
n2w_rcpp(par, hfun)
}
\arguments{
\item{par}{Parameters on their natural scale}

\item{hfun}{Name of hazard detection function}
}
\value{
Unconstrained parameters (scaled to the set of real numbers)
}
\description{
Scale parameters from their "natural" interval to their "working" interval,
the set of real numbers (for unconstrained optimization).
Written in C++.
}

