% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimation.R
\name{negllik.xandy}
\alias{negllik.xandy}
\title{HMM line transect model likelihood with perpendicular and forward distances.}
\usage{
negllik.xandy(b, xy, FUN, models = list(y = NULL, x = NULL), pm, Pi, delta, W,
  ymax, dy, nx = 100, groupfromy = NULL)
}
\arguments{
\item{b}{likelihood function parameters.}

\item{xy}{detections data, including x and y values for each detection, plus any covariates
in \code{model}).}

\item{FUN}{detection hazard function name.}

\item{models}{model specification, as in \code{\link{est.hmltm}}.}

\item{pm}{state-dependent Bernoulli parameters.}

\item{Pi}{Markov model transition probability matrix.}

\item{delta}{Markov model stationary distribution.}

\item{W}{perpendicular truncation distance for fitting.}

\item{ymax}{forward distance by which detection hazard has fallen to zero.}

\item{dy}{Markov model distance step size.}

\item{nx}{number of x-values to use in evaluating detection function.}

\item{groupfromy}{a forward distance (y) below which all y's are grouped into a single
interval in the likelihood function (i.e. exact y,s < groupfromy are combined into
an interval rather than passed as exact distances).}
}
\description{
Evaluates HMM line transect model likelihood with both perp. and forward distance data.
}

