% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pxy_double_rccp}
\alias{pxy_double_rccp}
\title{Probability of detection (double observer)}
\usage{
pxy_double_rccp(obs1, obs2, hfun, b1, b2, pcu, Pi, delta, ymax, dy, ally, cdf)
}
\arguments{
\item{hfun}{Hazard function name.}

\item{b1}{Hazard function parameter vector of first observer.}

\item{b2}{Hazard function parameter vector of second oberver.}

\item{pcu}{Bernoulli state-dependent probability parameters.}

\item{Pi}{Markov model transition probability matrix.}

\item{delta}{Markov model stationary distribution.}

\item{ymax}{Maximum forward distance in view.}

\item{dy}{Forward distance step increment.}

\item{ally}{Flag for whether or not to return probabilities at all forward distances in view.}

\item{cdf}{Flag for whether or not to return cumulative distribution function in forward dimension.
This differs from specifying ally=TRUE in that ally=TRUE calculates the cdf from ymax
to y=0, whereas cdf=TRUE calculates the cdf from ymax to y.}

\item{x}{Perpendicular distance.}

\item{y}{Forward distance.}
}
\description{
Calculate probability of detection at (x,y), or if cdf=TRUE, by (x,y), in the case
of a model with two observers.
Written in C++.
}

