% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pxy_simple_rcpp}
\alias{pxy_simple_rcpp}
\title{Probability of detection (single observer)}
\usage{
pxy_simple_rcpp(x, y, hfun, b, pcu, Pi, delta, ymax, dy, ally, cdf)
}
\arguments{
\item{x}{Perpendicular distance.}

\item{y}{Forward distance.}

\item{hfun}{Hazard function name.}

\item{b}{Hazard function parameter vector.}

\item{pcu}{Bernoulli state-dependent probability parameters.}

\item{Pi}{Markov model transition probability matrix.}

\item{delta}{Markov model stationary distribution.}

\item{ymax}{Maximum forward distance in view.}

\item{dy}{Forward distance step increment.}

\item{ally}{Flag for whether or not to return probabilities at all forward distances in view.}

\item{cdf}{Flag for whether or not to return cumulative distribution function in forward dimension.
This differs from specifying ally=TRUE in that ally=TRUE calculates the cdf from ymax 
to y=0, whereas cdf=TRUE calculates the cdf from ymax to y.}
}
\description{
Calculate probability of detection at (x,y), or if cdf=TRUE, by (x,y), in the case
of a model with one observer.
Written in C++.
}

