% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simData.R
\name{simData_double}
\alias{simData_double}
\title{Simulation (two observers)}
\usage{
simData_double(nbObs, pcu, gamma, b, hfun, xmax, ymax, ystep)
}
\arguments{
\item{nbObs}{Number of observations to simulate}

\item{pcu}{Vector of state dependent availability parameters (Pr(available|state))}

\item{gamma}{Transition probability matrix of the hidden state process}

\item{b}{List of vectors of parameters of the detection hazard function for each
observer}

\item{hfun}{Name of the detection hazard function (or vector of two names, if the
two observers have different detection functions)}

\item{xmax}{Maximum perpendicular distance in view}

\item{ymax}{Maximum forward distance in view}

\item{ystep}{Discretization step for the forward distance}
}
\value{
A data frame with the following columns:
\item{j}{Index of the observer}
\item{d}{1 if observed, 0 otherwise}
\item{x}{Perpendicular distance of the observation}
\item{y}{Forward distance of the observation}
}
\description{
Simulate detections in a hidden state line transect model with two observers.
}
\details{
If xmax and ymax are not specified, they are computed to be the points
in which h(xmax,0) (resp. h(0,ymax)) is approximately 0.01*h(0,0). If ystep is
not specified, the y-dimension is discretized to 100 intervals.
}
\examples{
pcu <- c(0.9,0)
gamma <- matrix(c(0.9,0.1,0.1,0.9),nrow=2)
b <- list(c(0.9,0.75,10),c(0.5,0.8,8))
hfun <- "h.EP2.0"
data <- simData_double(nbObs=100,pcu=pcu,gamma=gamma,b=b,hfun=hfun,xmax=100,ymax=100,ystep=1)
}

