% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simData_simple.R
\name{simData_simple}
\alias{simData_simple}
\title{Simulation (one observer)}
\usage{
simData_simple(nbAnimals, pcu, gamma, b, hfun, xmax, ymax, ystep)
}
\arguments{
\item{nbAnimals}{Number of animals}

\item{pcu}{Vector of state dependent availability parameters (Pr(available|state))}

\item{gamma}{Transition probability matrix of the hidden state process}

\item{b}{Vector of parameters of the detection hazard function}

\item{hfun}{Name of the detection hazard function}

\item{xmax}{Maximum perpendicular distance in view}

\item{ymax}{Maximum forward distance in view}

\item{ystep}{Discretization step for the forward distance}
}
\value{
A data frame of coordinates x and y of detected animals.
}
\description{
Simulate detections in a hidden state line transect model with one observer.
}
\examples{
pcu <- c(0.9,0)
gamma <- matrix(c(0.9,0.1,0.1,0.9),nrow=2)
b <- c(0.9,0.75,10)
hfun <- "h.EP2.0"
data <- simData_simple(nbAnimals=100,pcu=pcu,gamma=gamma,b=b,hfun=hfun,xmax=80,ymax=80,ystep=1)
}

