% Generated by roxygen2 (4.0.2): do not edit by hand
\name{truncdat}
\alias{truncdat}
\title{Distance data truncation.}
\usage{
truncdat(dat, minx = 0, maxx = NULL, twosit = FALSE,
  colnames = c("stratum", "area", "transect", "L", "x", "obs"))
}
\arguments{
\item{dat}{distance data frame. Must have columns for stratum stratum.area, transect, transect.length,
distance, and if \code{twosit}==TRUE then object as well. These can have any names, but the names
must be specified, via argument \code{colnames}.}

\item{minx}{perpendicular distance left-truncation point.}

\item{maxx}{perpendicular distance right-truncation point.}

\item{twosit}{If TRUE, assumes this is an mrds-type dataset with two lines per detection.}

\item{colnames}{name of columns containing stratum, stratum.area, transect, transect.length,
distance, and if \code{twosit}==TRUE then object as well, IN THIS ORDER in a character
vector. The default value is colnames=c("stratum","area","transect","L","x","obs").}
}
\description{
Truncates perp dists in data frame dat, inserting effort-only record if truncation removes
all detections on a transect.
Subtracts left-trunction point off all perp distances - after all truncation.
}

