% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{w2n_rcpp}
\alias{w2n_rcpp}
\title{Parameters inverse transformation}
\usage{
w2n_rcpp(par, hfun)
}
\arguments{
\item{par}{Unconstrained parameters (scaled to the set of real numbers)}

\item{hfun}{Name of hazard detection function}
}
\value{
Parameters on their natural scale
}
\description{
Scale parameters back from the set of real numbers ("working" interval)
to their "natural" interval.
Written in C++.
}

