\docType{data}
\name{bowhead.depths}
\alias{bowhead.depths}
\title{Greenland bowhead whale depth data time series.}
\format{A list of 8 numeric vectors. Their lengths are 191,048, 73,871, 24,673, 9,385, 37,689,
47,222, 15,490 and 39,989, respectively.}
\description{
  A list with elements \code{$a1} to \code{$a8}. Element
  \code{$ai} has the depths of tagged bowhead whale i each
  second.
}
\keyword{datasets}

