\name{hmmlt.gof.x}
\alias{hmmlt.gof.x}
\title{Goodness-of-fit in perpendicular dimension.}
\usage{
  hmmlt.gof.x(hmmlt, ks.plot = TRUE)
}
\arguments{
  \item{hmmlt.}{}

  \item{ks.plot}{If TRUE, does Q-Q plot.}
}
\description{
  Calculates goodness-of-fit in perpendicular dimension,
  plots fit, and returns p-value and other stuff.
}

