% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hmmlt.gof.x}
\alias{hmmlt.gof.x}
\title{Goodness-of-fit in perpendicular dimension.}
\usage{
hmmlt.gof.x(hmltm, ks.plot = TRUE)
}
\arguments{
\item{hmltm}{fitted model, as output by \code{\link{est.hmltm}}}

\item{ks.plot}{If TRUE, does Q-Q plot.}
}
\description{
Calculates goodness-of-fit in perpendicular dimension, plots fit, and returns p-value and other stuff.
}

