% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pxy_double_rccp}
\alias{pxy_double_rccp}
\title{Probability of detection (double observer)}
\usage{
pxy_double_rccp(x, y, d, hfun, b, pcu, Pi, delta, ymax, dy, ally, cdf)
}
\arguments{
\item{x}{Matrix of perpendicular distances (two columns: one for each observer)}

\item{y}{Matrix of forward distances (two columns: one for each observer)}

\item{d}{Matrix of detections (two columns: one for each observer). d[i,j] is
1 if observer j detected animal i, 0 otherwise.}

\item{hfun}{Hazard function name.}

\item{b}{Hazard function parameter matrix (two columns: one for each observer).}

\item{pcu}{Bernoulli state-dependent probability parameters.}

\item{Pi}{Markov model transition probability matrix.}

\item{delta}{Markov model stationary distribution.}

\item{ymax}{Maximum forward distance in view.}

\item{dy}{Forward distance step increment.}

\item{ally}{Flag for whether or not to return probabilities at all forward distances in view.}

\item{cdf}{Flag for whether or not to return cumulative distribution function in forward dimension.
This differs from specifying ally=TRUE in that ally=TRUE calculates the cdf from ymax 
to y=0, whereas cdf=TRUE calculates the cdf from ymax to y.}
}
\description{
Calculate probability of detection at (x,y), or if cdf=TRUE, by (x,y), in the case
of a model with two observers.
Written in C++.
}

